updated and changes for ball transformation
this one is to attach to a GameObject parent that contain both formations of "normal/regular" and "ball form". Think of it like this is fore ball transformation that need to be with the parent folder, and within this parent folder, there are two child that can only be active one at a time which is "normalForm" and "ballForm" and the parent folder is responsible for toggles the visibility of the two child based on the key input.